20.11.10 完善进货及退货管理

+++ src/main/java/com/lx/bus/mapper/OutportMapper.java
@@ -4,6 +4,5 @@ import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.lx.bus.domain.Outport;
 
 public interface OutportMapper extends BaseMapper<Outport> {
-
-    Integer queryOutPortSumByInportId(int inportid);
+    Integer queryOutPortSumByInportId(Long inportid);
 }
